lek1209

+++ Lek1209/src/Opgaver/opgave4_3.java
@@ -6,16 +6,16 @@ public class opgave4_3 {
     public static void main(String[] args) {
 
         System.out.print("Indtast tal: ");
-        int a = readInt();
+        double a = readDouble();
         System.out.print("Indtast tal: ");
-        int b = readInt();
+        double b = readDouble();
 
         System.out.println();
         System.out.println("Resultater:");
         System.out.println("Sum: " + (a+b));
         System.out.println("Difference: " + (a-b));
         System.out.println("Product: " + (a*b));
-        System.out.println("Average: " + (double)(a+b)/2);
+        System.out.println("Average: " + (a+b)/2);
         System.out.println("Absolute distance: " + Math.abs(a-b));
         System.out.println("Max: " + Math.max(a, b));
         System.out.println("Min: " + Math.min(a, b));
